CSCTTV-4139 TTV model update 2025-01 (#256)
* CSCTTV-4139 DimName.Id changed from int to long. Catchup all other changes in the DB context.

* Modify DimName.Id from int to long in source code

+++ aspnetcore/src/api/Models/Orcid/OrcidImportHelper.cs
@@ -19,7 +19,7 @@ namespace api.Models.Orcid
             dimProfileOnlyFundingDecisionIds = new();
         }
 
-        public List<int> dimNameIds { get; set; }
+        public List<long> dimNameIds { get; set; }
         public List<int> dimWebLinkIds { get; set; }
         public List<int> dimResearcherDescriptionIds { get; set; }
         public List<int> dimEmailAddressIds { get; set; }
